+++ lib/barking_iguana/compound/test_suite.rb
@@ -15,6 +15,17 @@ module BarkingIguana
             test.run
           end.add_description "Run #{test.name} test from #{name} suite"
 
+          test.stages.each do |stage|
+            Rake::Task.define_task "#{name}:#{test.name}:#{stage.name}" do
+              stage.run
+            end.add_description "Run #{stage.name} stage of the #{test.name} test from #{name} suite"
+            stage.actions.each do |action|
+              Rake::Task.define_task "#{name}:#{test.name}:#{stage.name}:#{action}" do
+                stage.public_send action
+              end.add_description "Run action #{action} for #{stage.name} stage of the #{test.name} test from #{name} suite"
+            end
+          end
+
           Rake::Task.define_task "#{name}:#{test.name}:destroy" do
             test.teardown
           end.add_description "Tear down #{test.name} test from #{name} suite"